// prodis-prodiswcl/src/app/modules/prodisapi/model/ccnl.ts
/*
* SPDX-FileCopyrightText: Copyright 2022 | Regione Piemonte
* SPDX-License-Identifier: EUPL-1.2
*/
/**
 * Prodis
 * API per il backend della suite di Prodis.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { DatiAzienda } from './datiAzienda';


export interface Ccnl { 
    ccnl?: Ccnl;
    ccnls?: Array<Ccnl>;
    codCcnlMin?: string;
    datiAziendas?: Array<DatiAzienda>;
    dsCcnl?: string;
    dtFine?: Date;
    dtInizio?: Date;
    dtTmst?: Date;
    id?: number;
    idNewCcnl?: number;
    settore?: string;
}